Brings in 8 upstream commits (TieruYT/metin2-playerbots): real offline
shops on the mt2009 line (an ikashop stand replaces the classic stall,
2.0.26), a core-crashing 64-bit-yang log-format bug (Teleporter refusal
SIGSEGV'd instead of logging - fixed in twelve places, all but one via
clean auto-merge), a weapon-scoring rewrite, and stall/world-reset fixes.

Resolved 4 conflicts: launcher/server-update-files.mt2009.txt (took
upstream's playerbot_* glob over our grown-by-hand enumeration - matches
CLAUDE.md's "never write another list" and self-covers every fork-only
file), playerbot_manager.cpp (kept both new includes -
playerbot_offline_market.h and our playerbot_language.h), playerbot_status.h
(kept the thin LLM-seam stub per the established pattern, ported upstream's
one real fix - the %lld gold format - into playerbot_llm_status.h instead
of adopting the ~570-line Polish rewrite), update-manifest-mt2009.json
(kept this fork's own v2.0.24 release URL rather than bump to upstream's
TieruYT-hosted v2.0.27 zip - a known, already-documented gap until this
fork publishes its own release).

Audited every upstream commit touching the playerbot overlay individually
(not just the conflicts) for new player-visible text: found none beyond
the one Teleporter line already ported. admin_panel.py's one real change
(a new offline-shop ranking query) needed no new strings; seban-panel had
zero changes in this range.

Verified: py_compile on the touched Python files, the existing and new
(tests/playerbot_offline_policy_test.cpp) unit tests both compile and pass
with plain g++, brace/paren balance on every hand-edited file, no leftover
conflict markers. Not verified: the real -fsyntax-only check against
m2src-cache/Docker, neither available in this sandbox - run by hand before
deploying. Full details in UPSTREAM_SYNC_2026-09-12.md.

+++ linux-port/overlays/playerbot/src/game/src/playerbot_activities.h
@@ -764,17 +764,17 @@ namespace
 				count = affordable;
 				price = GetPlayerBotNpcPurchasePrice(proto, count);
 				PlayerBotLogThrottled("tackle_part_buy", dwNow,
-						"PLAYERBOT_FISHING: buying what it can afford of %s pid=%u name=%s vnum=%u count=%d price=%lld gold=%d",
+						"PLAYERBOT_FISHING: buying what it can afford of %s pid=%u name=%s vnum=%u count=%d price=%lld gold=%lld",
 						what, ch->GetPlayerID(), ch->GetName(), vnum, count,
-						price, ch->GetGold());
+						price, (long long)ch->GetGold());
 			}
 		}
 		if (price <= 0 || ch->GetGold() < price)
 		{
 			PlayerBotLogThrottled("tackle_no_gold", dwNow,
-					"PLAYERBOT_FISHING: cannot afford %s pid=%u name=%s vnum=%u count=%d price=%lld gold=%d",
+					"PLAYERBOT_FISHING: cannot afford %s pid=%u name=%s vnum=%u count=%d price=%lld gold=%lld",
 					what, ch->GetPlayerID(), ch->GetName(), vnum, count,
-					price, ch->GetGold());
+					price, (long long)ch->GetGold());
 			return false;
 		}
 		// The bag has to have room BEFORE the purchase, because AutoGiveItem
@@ -799,8 +799,8 @@ namespace
 		if (!bMergesIntoStack && ch->GetEmptyInventory(size) < 0)
 		{
 			PlayerBotLogThrottled("tackle_no_room", dwNow,
-					"PLAYERBOT_FISHING: no bag room for %s pid=%u name=%s vnum=%u count=%d size=%d gold=%d",
-					what, ch->GetPlayerID(), ch->GetName(), vnum, count, size, ch->GetGold());
+					"PLAYERBOT_FISHING: no bag room for %s pid=%u name=%s vnum=%u count=%d size=%d gold=%lld",
+					what, ch->GetPlayerID(), ch->GetName(), vnum, count, size, (long long)ch->GetGold());
 			return false;
 		}
 		LPITEM bought = ch->AutoGiveItem(vnum, count, -1, false);
@@ -857,8 +857,8 @@ namespace
 			if (!pass)
 				return false;
 			PlayerBotChangeGold(ch, -(int)PLAYERBOT_FISHING_PASS_PRICE);
-			sys_log(0, "PLAYERBOT_FISHING: fishing pass bought pid=%u name=%s price=%u gold=%d",
-					ch->GetPlayerID(), ch->GetName(), PLAYERBOT_FISHING_PASS_PRICE, ch->GetGold());
+			sys_log(0, "PLAYERBOT_FISHING: fishing pass bought pid=%u name=%s price=%u gold=%lld",
+					ch->GetPlayerID(), ch->GetName(), PLAYERBOT_FISHING_PASS_PRICE, (long long)ch->GetGold());
 		}
 		if (!ch->EquipItem(pass))
 		{
